Apply new sizing behavior and refactor to reduce file size
